+++ config/fantom.ts
@@ -68,7 +68,7 @@ export default <NetworkData>{
         ],
     },
     rpcUrl: env.DRPC_API_KEY
-        ? `https://lb.drpc.org/ogrpc?network=fantom&dkey=${env.DRPC_API_KEY}`
+        ? `https://lb.drpc.org/ogrpc?network=fantom&dkey=${env.DRPC_BEETS_API_KEY}`
         : `https://rpc.ankr.com/fantom`,
     rpcMaxBlockRange: 1000,
     protocolToken: 'beets',